+++ utils/types.ts
@@ -1,5 +1,6 @@
-export type DeepPartial<T> = T extends Record<string, any>
-  ? {
-      [K in keyof T]?: DeepPartial<T[K]>;
-    }
-  : T;
+export type DeepPartial<T> =
+  T extends Record<string, any>
+    ? {
+        [K in keyof T]?: DeepPartial<T[K]>;
+      }
+    : T;